+++ packages/scripts/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Feature
 
-- Add `--alpha` option to use latest alpha in generator @sneridagh
+- Add `--canary` option to use latest alpha in generator @sneridagh
 
 ### Bugfix